debug: Add visible alerts to test click functionality on deployed site
Adding console.log when component loads and alert() when cards are clicked
to definitively test if the JavaScript is working on the deployed version.
This will help us determine if the issue is:
1. Component not loading at all
2. Click handlers not working
3. Modal state management issue

If you see the alert when clicking cards, we know JS is working.
If no alert, then there's a fundamental JS loading issue.